+++ src/QuickMarcEditor/utils.test.js
@@ -612,6 +612,64 @@ describe('QuickMarcEditor utils', () => {
     });
   });
 
+  describe('checkControlFieldLength', () => {
+    it('should not return error message when only one 001 field is present', () => {
+      const formValues = {
+        records: [{
+          tag: '001',
+          content: 'some content',
+          id: 'id001',
+        }, {
+          tag: '035',
+          content: 'some content',
+          id: 'id035',
+          indicators: ['0', '\\'],
+        }, {
+          content: '',
+          id: 'id999ff',
+          indicators: ['f', 'f'],
+          tag: '999',
+        }],
+        updateInfo: {
+          recordState: 'actual',
+          updateDate: '01/01/1970',
+        },
+      };
+
+      expect(utils.checkControlFieldLength(formValues)).not.toBeDefined();
+    });
+
+    it('should return error message when more than one 001 field is present', () => {
+      const formValues = {
+        records: [{
+          tag: '001',
+          content: 'some content',
+          id: 'id001-1',
+        }, {
+          tag: '001',
+          content: 'some other content',
+          id: 'id001-2',
+        }, {
+          tag: '035',
+          content: 'some content',
+          id: 'id035',
+          indicators: ['0', '\\'],
+        }, {
+          content: '',
+          id: 'id999ff-746e-4058-a35b-8130e4f6d277',
+          indicators: ['f', 'f'],
+          tag: '999',
+        }],
+        updateInfo: {
+          recordState: 'actual',
+          updateDate: '01/01/1970',
+        },
+      };
+
+      expect(utils.checkControlFieldLength(formValues).props.id).toBe('ui-quick-marc.record.error.controlField.multiple');
+    });
+  });
+
   describe('validateSubfield', () => {
     it('should not return error message when indicators are present and content is not empty', () => {
       const initialRecords = [